refactor(models): move ml_models.py -> backend/ml/runtime_models.py
backend/models/ml_models.py held PyTorch/XGBoost/LightGBM/Prophet runtime
classes (LSTMModel, TransformerModel, LightGBMModel, ModelManager,
PredictionResult, StockDataset). These are not ORM schema; they belong
under backend/ml/ alongside ml_tables.py and training_pipeline.py.

Renamed to backend/ml/runtime_models.py (not ml_models.py to avoid
collision with the existing ml_tables.py naming and to make the runtime
nature explicit). Recorded as a git rename so history follows.

Rewrote 5 callers:
  scripts/train_ml_models.py
  scripts/load_trained_models.py
  backend/analytics/recommendation_types.py
  backend/analytics/recommendation_optimized.py
  backend/analytics/recommendation_engine.py

Smoke imports of {ModelManager, PredictionResult, LSTMModel, TransformerModel}
and the three analytics callers all return OK.

Part of Tier B1 (docs/ANALYSIS_2026-05-14.md).
